Bug 981510 - Define MVVM as AMD modules
