+++ apps/settings/test/unit/modules/mvvm/list_view_test.js
@@ -1,34 +1,69 @@
-/* global ListView:false */
-/* global ObservableArray:false */
+/* global define */
 'use strict';
 
 // The views are very tightly coupled to models, rather than reimplement model
 // behavior in a mock, rely on tested interfaces on the model instead
-requireApp('settings/js/mvvm/models.js');
-requireApp('settings/js/mvvm/views.js');
-
 suite('ListView', function() {
+  // Define a shim for ObservableArray that is to be loaded in ListView.
+  // Then we are able to creat a stub on it.
 
-  var suiteSandbox = sinon.sandbox.create();
-  suiteSetup(function() {
-    var OriginalObservableArray = ObservableArray;
-    // stub the real ObservableArray to create spies on observe method
-    suiteSandbox.stub(window, 'ObservableArray', function() {
-      var result = OriginalObservableArray.apply(this, arguments);
-      sinon.spy(result, 'observe');
-      sinon.spy(result, 'unobserve');
-      return result;
+  suiteSetup(function(done) {
+    // Use the shim in ListView.
+    var ObservableArrayShimName = 'observable_array_shim' + Date.now();
+    var map = {
+      'modules/mvvm/list_view': {
+        'modules/mvvm/observable_array': ObservableArrayShimName
+      }
+    };
+
+    // XXX: Create a new require context.
+    //      We can only define the mock module after retrieving the context.
+    var requireCtx = testRequire([], map, function() {});
+    // Define the mock module.
+    define(ObservableArrayShimName, function() {
+      var ctor = function observable_array_shim() {
+        return ctor.mInnerFunction.apply(this, arguments);
+      };
+      ctor.mInnerFunction = function() {};
+      ctor.mTeardown = function observable_array_shim_teardown() {
+        ctor.mInnerFunction = function() {};
+      };
+      return ctor;
     });
 
+    requireCtx([
+      'modules/mvvm/observable',
+      'modules/mvvm/observable_array',
+      'modules/mvvm/list_view',
+      ObservableArrayShimName
+    ], (function(Observable, ObservableArray, ListView, ObservableArrayShim) {
+      this.Observable = Observable;
+      this.ListView = ListView;
+      this.ObservableArrayShim = ObservableArrayShim;
+      // stub the real ObservableArray to create spies on observe method
+      // Redirect the shim to the ObservableArray stub.
+      this.ObservableArray = sinon.stub(ObservableArrayShim,
+        'mInnerFunction', function() {
+          var result = ObservableArray.apply(this, arguments);
+          sinon.spy(result, 'observe');
+          sinon.spy(result, 'unobserve');
+          return result;
+      });
+      done();
+    }).bind(this));
   });
+
+  teardown(function() {
+    this.ObservableArray.reset();
+  });
+
   suiteTeardown(function() {
-    suiteSandbox.restore();
+    this.ObservableArray.restore();
   });
 
   setup(function() {
     this.array = [];
-    this.observableArray = ObservableArray(this.array);
-    ObservableArray.reset();
+    this.observableArray = this.ObservableArray(this.array);
   });
 
   suite('ListView(<ul>, ObservableArray, function)', function() {
@@ -41,7 +76,7 @@ suite('ListView', function() {
         return document.createElement('li');
       }).bind(this);
       this.sinon.spy(this, 'template');
-      this.listView = ListView(
+      this.listView = this.ListView(
         this.container, this.observableArray, this.template
       );
     });
@@ -242,7 +277,7 @@ suite('ListView', function() {
         });
 
         test('creates and watches an ObservableArray', function() {
-          var withNew = ObservableArray.withArgs(this.array);
+          var withNew = this.ObservableArray.withArgs(this.array);
           assert.ok(withNew.called);
           var observe = withNew.returnValues[0].observe;
           assert.ok(observe.calledWith('insert'));
@@ -345,7 +380,7 @@ suite('ListView', function() {
       }).bind(this);
       this.sinon.spy(this, 'template');
       this.observableArray.push({ count: 0 });
-      this.listView = ListView(
+      this.listView = this.ListView(
         this.container, this.observableArray, this.template
       );
     });
@@ -472,8 +507,8 @@ suite('ListView', function() {
 
   suite('observable hooks', function() {
     setup(function() {
-      this.observable = Observable({ test: 1 });
-      this.observableArray = ObservableArray([this.observable]);
+      this.observable = this.Observable({ test: 1 });
+      this.observableArray = this.ObservableArray([this.observable]);
       this.hook = this.sinon.spy();
       this.template = (function(data, recycle, helper) {
         helper.observeAndCall(data, {
@@ -483,7 +518,7 @@ suite('ListView', function() {
       }).bind(this);
       this.container = document.createElement('ul');
       this.sinon.spy(this, 'template');
-      this.listView = ListView(
+      this.listView = this.ListView(
         this.container, this.observableArray, this.template
       );
     });
@@ -521,9 +556,9 @@ suite('ListView', function() {
 
   suite('destroy', function() {
     setup(function() {
-      this.observableArray = ObservableArray([1, 2, 3]);
+      this.observableArray = this.ObservableArray([1, 2, 3]);
       this.container = document.createElement('div');
-      this.listView = ListView(
+      this.listView = this.ListView(
         this.container, this.observableArray, function() {
           return document.createElement('div');
         }
@@ -557,7 +592,7 @@ suite('ListView', function() {
       this.observableArray.observe.reset();
       this.sinon.spy(this.listView, 'destroy');
       // create a new one, we don't care about what happens to it.
-      ListView(
+      this.ListView(
         this.container, this.observableArray, function() {
           return document.createElement('div');
         }
@@ -578,4 +613,3 @@ suite('ListView', function() {
     });
   });
 });
-
